sprint 6 version 1

+++ src/Subtask.java
@@ -1,13 +1,13 @@
 import java.util.ArrayList;
 
 public class Subtask extends Task{
-    private int epicId;
+    private long epicId;
     private Epic epic;
     private ArrayList<Long> subtaskIds = new ArrayList<>();
 
     public Subtask(String name, String description, long epicId) {
         super(name, description);
-        this.epicId = (int) epicId;
+        this.epicId = (long) epicId;
     }
 
     public void addSubtaskId(long subtaskId) {
@@ -27,7 +27,7 @@ public class Subtask extends Task{
         return epicId;
     }
 
-    public void setEpicId(int epicId) {
+    public void setEpicId(long epicId) {
         if (epicId == this.getId()) {
             throw new IllegalArgumentException("ошибка");
         }
@@ -43,6 +43,4 @@ public class Subtask extends Task{
                 ", status=" + getStatus() +
                 '}';
     }
-
-
 }
